Moved components to components folder, changed login axios calls to match backend routes
