+++ client/src/components/Login/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { login } from '../redux/actions/auth';
+import { login } from '../../redux/actions/auth';
 
 class Login extends React.Component {
   state = { pin: '', pass: '' };
@@ -20,19 +20,24 @@ class Login extends React.Component {
       <div>
         <Link to="/register">Don't have an account? Register here</Link>
         <form onSubmit={this.handleSubmit}>
-          <input
+        <input
             placeholder="pin"
             type="text"
             onChange={this.handleChange}
             name="pin"
+            minLength="6"
+            maxLength="6"
+            autoComplete="username"
             value={this.state.pin}
           />
           <input
             placeholder="password"
             type="password"
-            minLength="8"
             onChange={this.handleChange}
             name="pass"
+            minLength="8"
+            maxLength="30"
+            autoComplete="current-password"
             value={this.state.pass}
           />
           <button type="submit">Submit</button>
