+++ client/src/components/NotFound/index.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+class NotFound extends Component {
+  render() {
+    return (
+      <div>404</div>
+    );
+  }
+}
+
+export default NotFound;
