+++ client/src/components/Register/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { register } from '../redux/actions/auth';
+import { register } from '../../redux/actions/auth';
 
 class Register extends React.Component {
   state = { firstName: '', lastName: '', pin: '', pass: '', confirmPass: '' };
@@ -24,6 +24,7 @@ class Register extends React.Component {
               type="text"
               onChange={this.handleChange}
               name="firstName"
+              maxLength="30"
               value={this.state.firstName}
             />
             <input
@@ -31,6 +32,7 @@ class Register extends React.Component {
               type="text"
               onChange={this.handleChange}
               name="lastName"
+              maxLength="30"
               value={this.state.lastName}
             />
             <input
@@ -38,12 +40,15 @@ class Register extends React.Component {
               type="text"
               onChange={this.handleChange}
               name="pin"
+              minLength="6"
+              maxLength="6"
               value={this.state.pin}
             />
             <input
               placeholder="password"
               type="password"
               minLength="8"
+              maxLength="30"
               onChange={this.handleChange}
               name="pass"
               value={this.state.pass}
@@ -52,6 +57,7 @@ class Register extends React.Component {
               placeholder="confirm password"
               type="password"
               minLength="8"
+              maxLength="30"
               onChange={this.handleChange}
               name="confirmPass"
               value={this.state.confirmPass}
