+++ client/src/redux/actions/auth.js
@@ -13,21 +13,20 @@ export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
 // Axios Defaults
 
 axios.defaults.withCredentials = true;
-axios.defaults.headers.common['Authorization'] =
-  'Bearer ' + localStorage.getItem('jwt');
+axios.defaults.headers.common['Authorization'] = localStorage.getItem('jwt');
 
 // Actions
 
 export const login = ({ pin, pass }, push) => {
   return (dispatch) => {
     axios
-      .post(`${serverURI}/login`, { pin, pass })
+      .post(`${serverURI}/api/employees/login`, { pin, pass })
       .then((res) => {
-        const { role } = jwt_decode(res.data);
+        const { role } = jwt_decode(res.data.token);
 
-        dispatch({ type: LOGIN_SUCCESS, jwt: res.data, role });
+        dispatch({ type: LOGIN_SUCCESS, jwt: res.data.token, role });
 
-        localStorage.setItem("jwt", res.data)
+        localStorage.setItem("jwt", res.data.token);
 
         if (role.admin || role.manager) {
           push('/servers');
@@ -49,7 +48,7 @@ export const register = ({ firstName, lastName, pin, pass, confirmPass }) => {
     }
     dispatch({ type: PASSWORD_MATCH_SUCCESS });
     axios
-      .post(`${serverURI}/login`, { name: `${firstName} ${lastName}`, pin, pass })
+      .post(`${serverURI}/api/employees/register`, { name: `${firstName} ${lastName}`, pin, pass })
       .then((res) => {
         dispatch({ type: LOGIN_SUCCESS, payload: res.data });
       })
